common_drivers:xcvr_core:xcvr_modules CPLL lane rate calculation bug
The Linux driver calculates lane rate in Hz and at the end converts it
in kHz to represent it more compactly. This feature carried over at the
update of the no-OS driver, but this on calculates the rate directly in
kHz.

Remove the last division by 1000 to get the desired rate in the no-OS
driver.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>